[TodoList] feat: add global exception handel

+++ src/main/java/com/todo/todolist/service/TodoService.java
@@ -1,10 +1,12 @@
 package com.todo.todolist.service;
 
 import com.todo.todolist.entity.Todo;
+import com.todo.todolist.exception.NotFoundException;
 import com.todo.todolist.repository.TodoRepository;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Objects;
 
 @Service
 public class TodoService {
@@ -20,6 +22,10 @@ public class TodoService {
     }
 
     public void deleteTodo(Integer id) {
+        Todo existTodo = todoRepository.findById(id).orElse(null);
+        if (Objects.isNull(existTodo)) {
+            throw new NotFoundException();
+        }
         todoRepository.deleteById(id);
     }
 
@@ -28,6 +34,18 @@ public class TodoService {
     }
 
     public Todo updateTodo(Integer id, Todo todo) {
+        Todo existTodo = todoRepository.findById(id).orElse(null);
+        if (Objects.isNull(existTodo)) {
+            throw new NotFoundException();
+        }
         return todoRepository.save(todo);
+    }
+
+    public Todo getTodoById(Integer id) {
+        Todo existTodo = todoRepository.findById(id).orElse(null);
+        if (Objects.isNull(existTodo)) {
+            throw new NotFoundException();
+        }
+        return existTodo;
     }
 }
